--- office-lunch-menu-management/backend/controllers/logout.js
@@ -1,13 +1,0 @@
-const blacklist = new Set();
-
-const logout = (req, res) => {
-  const token = req.header('Authorization').replace('Bearer ', '');
-  blacklist.add(token);
-  res.json({ message: 'Logged out successfully' });
-};
-
-const isTokenBlacklisted = (token) => {
-  return blacklist.has(token);
-};
-
-module.exports = { logout, isTokenBlacklisted };
