+++ office-lunch-menu-management/backend/routes/authRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
-const { register, login } = require('../controllers/authController');
-const { logout } = require('../controllers/logout');
+const { register, login, logout } = require('../controllers/authController');
+
 const router = express.Router();
 
 router.post('/register', register);
